+++ client/lib/controller.cpp
@@ -3,7 +3,7 @@
 // Created Date: 13/05/2016
 // Author: Seki Inoue
 // -----
-// Last Modified: 19/05/2020
+// Last Modified: 09/06/2020
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2016-2020 Hapis Lab. All rights reserved.
@@ -24,12 +24,10 @@
 
 #include "ec_config.hpp"
 #include "emulator_link.hpp"
-#include "ethercat_link.hpp"
 #include "firmware_version.hpp"
 #include "geometry.hpp"
 #include "link.hpp"
 #include "privdef.hpp"
-#include "soem_link.hpp"
 #include "timer.hpp"
 
 namespace autd {